pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sentry-mcp"
version = "0.1.0"
description = "A command-line tool for monitoring and analyzing Sentry metrics using MCP"
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
authors = [
    { name = "Kirtiranjan Sahoo", email = "skirtiranjan.sahoo@gmail.com" }
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: System :: Monitoring",
    "Topic :: System :: Systems Administration",
]
dependencies = [
    "requests==2.32.0",
    "litellm==1.75.0", 
    "anyio==4.5.0",
    "mcp-server==0.1.4",
    "python-dotenv==1.0.1",
    "backoff==2.2.0"
]

[project.scripts]
sentry-mcp = "sentry_mcp.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "build>=1.0.0",
    "twine>=4.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/sentry-mcp"
Documentation = "https://sentry-mcp.readthedocs.io/"
Repository = "https://github.com/yourusername/sentry-mcp.git"
Issues = "https://github.com/yourusername/sentry-mcp/issues"

[tool.black]
line-length = 88
target-version = ["py38"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 88

[tool.ruff]
line-length = 88
target-version = "py38"
select = ["E", "F", "B", "I", "N", "UP", "PL", "RUF"]
ignore = []

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_optional = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=sentry_mcp --cov-report=term-missing"
asyncio_mode = "auto"

[tool.hatch.build.targets.wheel]
packages = ["sentry_mcp"]
